fix ci test failures
