+++ chalice/local.py
@@ -5,7 +5,6 @@ This is intended only for local development purposes.
 """
 from __future__ import print_function
 from __future__ import annotations
-import datetime
 import re
 import threading
 import time
@@ -54,6 +53,7 @@ from chalice.app import WebsocketEvent  # noqa
 from chalice.app import AuthResponse  # noqa
 from chalice.app import BuiltinAuthConfig  # noqa
 from chalice.config import Config  # noqa
+from chalice.websocket import WebsocketClient
 
 from chalice.compat import urlparse, parse_qs
 
@@ -693,99 +693,6 @@ class ChaliceRequestHandler(BaseHTTPRequestHandler):
         self.end_headers()
 
 
-class WebsocketClientConnection:
-    def __init__(self, connection: WebsocketConnection) -> None:
-        self.connection = connection
-        self._connected_at = datetime.datetime.utcnow()
-        self._last_active_at = self._connected_at
-
-    def _touch(self) -> None:
-        self._last_active_at = datetime.datetime.utcnow()
-
-    def recv(self) -> str | bytes:
-        message = self.connection.recv()
-        self._touch()
-        return message
-
-    def send(self, message: str | bytes) -> None:
-        self.connection.send(message)
-        self._touch()
-
-    def close(self) -> None:
-        self.connection.close()
-        self._touch()
-
-    def info(self) -> Dict[str, Any]:
-        try:
-            source_ip = self.connection.remote_address[0]
-        except Exception:
-            source_ip = ''
-        if self.connection.request is not None:
-            try:
-                user_agent = next(iter(
-                    self.connection.request.headers.get_all('User-Agent')))
-            except StopIteration:
-                user_agent = ''
-        else:
-            user_agent = ''
-        return {
-            'ConnectedAt': self._connected_at,
-            'Identity': {
-                'SourceIp': source_ip,
-                'UserAgent': user_agent,
-            },
-            'LastActiveAt': self._last_active_at,
-        }
-
-
-class WebsocketClientExceptions:
-    GoneException = Exception
-
-
-class WebsocketClient:
-    exceptions = WebsocketClientExceptions()
-
-    def __init__(self) -> None:
-        self._connections: Dict[str, WebsocketClientConnection] = {}
-
-    def _get(self, connection_id) -> WebsocketClientConnection:
-        try:
-            return self._connections[connection_id]
-        except KeyError:
-            raise self.exceptions.GoneException('Connection not found')
-
-    def _del(self, connection_id: str):
-        try:
-            del self._connections[connection_id]
-        except KeyError:
-            raise self.exceptions.GoneException('Connection not found')
-
-    def get_connection_id(self, connection: WebsocketConnection) -> str:
-        return base64.b64encode(connection.id.bytes).decode('ascii')
-
-    def add_connection(self, connection: WebsocketConnection):
-        self._connections[self.get_connection_id(connection)] = (
-            WebsocketClientConnection(connection)
-        )
-
-    def receive_message(self, ConnectionId: str) -> str | bytes:
-        return self._get(ConnectionId).recv()
-
-    def post_to_connection(self, ConnectionId: str, Data: str) -> None:
-        try:
-            self._get(ConnectionId).send(Data)
-        except WebsocketConnectionClosed:
-            self._del(ConnectionId)
-            raise self.exceptions.GoneException('Connection closed')
-
-    def delete_connection(self, ConnectionId: str) -> None:
-        self._get(ConnectionId).close()
-        self._del(ConnectionId)
-
-    def get_connection(self, ConnectionId: str) -> Dict[str, Any]:
-        return self._get(ConnectionId).info()
-
-
 class ChaliceWsHandler:
     MAX_LAMBDA_EXECUTION_TIME = 900
 
@@ -798,6 +705,7 @@ class ChaliceWsHandler:
 
     @property
     def _websocket_client(self) -> WebsocketClient:
+        # pylint: disable=protected-access
         if isinstance(self.app_obj.websocket_api._client, WebsocketClient):
             return self.app_obj.websocket_api._client
         else:
@@ -878,7 +786,7 @@ class ChaliceWsHandler:
         return handler.handler_function(WebsocketEvent(
             event, self._generate_lambda_context()))
 
-    def __call__(self, websocket: WebsocketConnection):
+    def __call__(self, websocket: WebsocketConnection) -> None:
         connection_id = self._websocket_client.get_connection_id(websocket)
         try:
             while True:
@@ -977,6 +885,7 @@ class LocalDevServer(object):
         self.http_server.handle_request()
 
     def serve_forever(self) -> None:
+        # pylint: disable=protected-access
         if 'WEBSOCKETS' in self.app_object._features_used:
             print("Serving on http://%s:%s and ws://%s:%s" %
                   (self.host, self.port, self.ws_host, self.ws_port))
