+++ tests/functional/cli/test_factory.py
@@ -215,7 +215,9 @@ def test_error_raised_on_invalid_config_json(clifactory):
 def test_can_create_local_server(clifactory):
     app = clifactory.load_chalice_app()
     config = clifactory.create_config_obj()
-    server = clifactory.create_local_server(app, config, '0.0.0.0', 8000)
+    server = clifactory.create_local_server(
+        app, config, '0.0.0.0', 8000, '0.0.0.0', 8001
+    )
     assert isinstance(server, local.LocalDevServer)
     assert server.host == '0.0.0.0'
     assert server.port == 8000
