+++ tests/unit/cli/test_cli.py
@@ -17,7 +17,10 @@ def test_cannot_run_local_mode_with_trailing_slash_route():
     local_server = mock.Mock(spec=LocalDevServer)
     factory.create_local_server.return_value = local_server
     with pytest.raises(ValueError) as e:
-        cli.run_local_server(factory, 'localhost', 8000, local_stage_test)
+        cli.run_local_server(
+            factory, 'localhost', 8000, local_stage_test,
+            'localhost', 8001
+        )
     assert str(e.value) == 'Route cannot end with a trailing slash: foobar/'
