+++ tests/unit/test_local.py
@@ -726,20 +726,26 @@ def test_can_create_lambda_event_for_post_with_formencoded_body():
 
 
 def test_can_provide_port_to_local_server(sample_app):
-    dev_server = local.create_local_server(sample_app, None, '127.0.0.1',
-                                           port=23456)
+    dev_server = local.create_local_server(
+        sample_app, Config(), '127.0.0.1',
+        port=23456, ws_host='127.0.0.1', ws_port=23457
+    )
     assert dev_server.http_server.server_port == 23456
 
 
 def test_can_provide_host_to_local_server(sample_app):
-    dev_server = local.create_local_server(sample_app, None, host='0.0.0.0',
-                                           port=23456)
+    dev_server = local.create_local_server(
+        sample_app, Config(), '0.0.0.0',
+        port=23456, ws_host='0.0.0.0', ws_port=23457
+    )
     assert dev_server.host == '0.0.0.0'
 
 
 def test_wraps_custom_sample_app_with_local_chalice(custom_sample_app):
-    dev_server = local.create_local_server(custom_sample_app, None,
-                                           host='0.0.0.0', port=23456)
+    dev_server = local.create_local_server(
+        custom_sample_app, Config(), '0.0.0.0',
+        port=23456, ws_host='127.0.0.1', ws_port=23457
+    )
     assert isinstance(dev_server.app_object, local.LocalChalice)
     assert isinstance(dev_server.app_object, custom_sample_app.__class__)
     assert dev_server.app_object.custom_method() == 'foo'
